% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{Add index columns to a table}
\usage{
add_index(tab)
}
\arguments{
\item{tab}{Table generated from \code{\link{generate_table}} .}
}
\value{
Tibble as per \code{\link{generate_table}}  with
additional columns for indexes
}
\description{
Index is calculated by dividing each column by the Total column and
multiplying by 100.
}
