% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{calculate_means}
\alias{calculate_means}
\title{Calculate weighted means}
\usage{
calculate_means(
  input_data,
  col_var,
  respid_var,
  weight_var,
  var_labels,
  val_labels
)
}
\description{
Calculate weighted means
}
