% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVariable.R
\name{filterVariable}
\alias{filterVariable}
\alias{filterVariable_ui}
\alias{filterVariable_server}
\title{Shiny module for specifying respondent filter variables}
\usage{
filterVariable_ui(id)

filterVariable_server(id, varLabels, tab_data)
}
\arguments{
\item{id}{Module ID}

\item{varLabels}{Dataframe (e.g. as generated by
\link[tveDataLoader]{get_varLabels}) containing two columns:
\itemize{
\item \code{variable}: the variable names
\item \code{label}: the variable labels
}}

\item{tab_data}{Response data to be tabled.}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: a list with 3 reactive elements:
\itemize{
\item \code{inc_filters}: TRUE/FALSE. Status of the input switch for whether
filters should be included
\item \code{filter_vars}: A named character vector of the variables to be used as
respondent filters. To be passed to \link[datamods]{filter_data_server}.
\item \code{filter_dat}: A dataframe including the filter variables converted to
factors.  To be passed to \link[datamods]{filter_data_server}.
}
}
}
\description{
Generates an input switch for whether respondent filters should
be included or not, and if yes, a picker input that lists all variables from
which the filter variables should be chosen. Multiple filter variables can
be chosen.
}
