% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{generate_table}
\alias{generate_table}
\title{wraps the 3 generate table functions into one to produce final table}
\usage{
generate_table(
  dat,
  row_vars,
  col_var,
  weight_var,
  variable_labels = NULL,
  value_labels = NULL
)
}
\arguments{
\item{dat}{data}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{col_var}{Character string containing the variables to be used as
columns in the output table.}

\item{weight_var}{Character string containing the variable name of the weight
to be used.}

\item{variable_labels}{Either \code{NULL} (default) or a dataframe consisting of 2
columns:
\enumerate{
\item variable - variable names from \code{dat}
\item label - variable labels from \code{dat}
}

e.g. as generated by \code{\link[tveDataLoader:get_varLabels]{tveDataLoader::get_varLabels()}}. If \code{NULL}
then existing variable labels will be read from \code{dat} using
\code{\link[tveDataLoader:get_varLabels]{tveDataLoader::get_varLabels()}}. If a dataframe is provided then those
labels will be used to replace any existing labels via
\code{\link[tveDataLoader:update_var_labels]{tveDataLoader::update_var_labels()}}. Only the labels of variables provided
will be updated, others will remain unchanged and will be used in the table
as they are. As we're updating labels, every variable listed must
correspond to a column in \code{dat}.}

\item{value_labels}{Either \code{NULL} (default) or a dataframe consisting of 3
columns:
\enumerate{
\item variable - variable names from \code{dat}
\item value - values of the variables
\item value label - value labels of the values
}

e.g. as generated by \code{\link[tveDataLoader:get_valLabels]{tveDataLoader::get_valLabels()}}. If \code{NULL} then the
existing value labels will be read from \code{dat} using
\code{\link[tveDataLoader:get_valLabels]{tveDataLoader::get_valLabels()}}. If a dataframe is provided then those
value labels will be used to replace any existing labels via
\code{\link[tveDataLoader:update_val_labels]{tveDataLoader::update_val_labels()}}. Only the value labels of variables
provided will be updated, others will remain unchanged and used in the
table as they are. As we're updating labels, every variable listed must
correspond to a column in \code{dat}.}
}
\value{
final output table
}
\description{
wraps the 3 generate table functions into one to produce final table
}
\examples{
''

}
