% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table_unlabelled.R
\name{generate_table_unlabs}
\alias{generate_table_unlabs}
\title{wraps the 3 generate table functions into one to produce final table for unlabelled data}
\usage{
generate_table_unlabs(dat, row_vars, col_var, weight_var = NULL)
}
\arguments{
\item{dat}{data}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{col_var}{Character string containing the variables to be used as
columns in the output table.}

\item{weight_var}{Character string containing the variable name of the weight
to be used.}
}
\value{
final output table
}
\description{
wraps the 3 generate table functions into one to produce final table for unlabelled data
}
\examples{
''


}
