% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariableLabels.R
\name{getVariableLabels}
\alias{getVariableLabels}
\title{Shiny module to extract variable labels from data file}
\usage{
getVariableLabels(id, dat, exclude = "no_weighting")
}
\arguments{
\item{id}{Module ID}

\item{dat}{Data frame from which to read variable labels}

\item{exclude}{Character string of variables to exclude}
}
\description{
Takes a data file loaded in a shiny app and reads the variable labels. If a
variable has no labels, then it assigns the variable name as label. Any
variables supplied to \code{exclude} are excluded.
}
