% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{identify_var_type}
\alias{identify_var_type}
\title{identify variable type of input variables}
\usage{
identify_var_type(dat, row_vars, variable_labels, value_labels)
}
\arguments{
\item{dat}{data}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{variable_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item label - the variable labels of \code{dat}
As generated by \link[tveDataLoader]{get_varLabels}
}}

\item{value_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item value - the values \code{dat}
\item value label - the value labels of \code{dat}
As generated by \link[tveDataLoader]{get_valLabels}#'
}}
}
\value{
named list containing flags for multi code, single code and numerics
}
\description{
identify variable type of input variables
}
\examples{
''

}
