% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table_unlabelled.R
\name{identify_var_type_unlabs}
\alias{identify_var_type_unlabs}
\title{identify variable type of input variables for unlabelled data}
\usage{
identify_var_type_unlabs(dat, row_vars)
}
\arguments{
\item{dat}{data}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}
}
\value{
named list containing flags for numerics and categories
}
\description{
identify variable type of input variables for unlabelled data
}
\examples{
''


}
