% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{mean_calcs}
\alias{mean_calcs}
\title{calculate table for multi code and numeric questions}
\usage{
mean_calcs(dat, row_vars, col_var, weight_var, variable_labels, flag_list)
}
\arguments{
\item{dat}{data}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{col_var}{Character string containing the variables to be used as
columns in the output table.}

\item{weight_var}{Character string containing the variable name of the
weight to be used.}

\item{variable_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item label - the variable labels of \code{dat}
As generated by \link[tveDataLoader]{get_varLabels}
}}

\item{flag_list}{named list of character vectors for single code, single code and multi code
As generated by \link[tveTableTool]{identify_var_type}}
}
\value{
table
}
\description{
calculate table for multi code and numeric questions
}
\examples{
''
}
