% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputTab.R
\name{outputTab}
\alias{outputTab}
\alias{outputTab_ui}
\alias{outputTab_server}
\title{Shiny module for generating output table}
\usage{
outputTab_ui(id)

outputTab_server(
  id,
  dat,
  filtering_exp,
  dat_file_name,
  weight_var,
  row_var,
  col_var,
  variable_labels
)
}
\arguments{
\item{id}{Module ID}

\item{dat}{Datatable to be summarised. See \link{generate_table} for
requirements.}

\item{filtering_exp}{Description of filters to be incorporated into the
output table's caption.}

\item{dat_file_name}{Name of the SPSS source file for inclusion in the
output table's caption.}

\item{weight_var}{Weighting variable to be applied. "No_weighting" if no
weighting to be applied.}

\item{row_var}{Variables from \code{dat} to be used as rows in the output table.
Intended to be generated by \link{selectionVars} but can also be a
character vector of variables names from \code{dat}.}

\item{col_var}{Variable from \code{dat} to be used as column breaks. Intended to
be generated by \link{selectionVars} but can also be a character vector of a
variable name from \code{dat}.}

\item{variable_labels}{Dataframe (e.g. as generated by
\link[tveDataLoader]{get_varLabels}) containing two columns:
\itemize{
\item \code{variable}: the variable names
\item \code{label}: the variable labels
}}
}
\description{
Produces an output table using \link{generate_table}, then
formats and renders as a DT data table (\link[DT]{renderDataTable}).
}
