% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare data loaded by \link[datamods]{import-file}}
\usage{
prepData(id, data)
}
\arguments{
\item{id}{Module ID}

\item{data}{Dataframe loaded by \link[datamods]{import-file}}
}
\value{
Dataframe with only numeric variables and no_weighting variable.
}
\description{
Shiny module to process .sav files loaded via \link[datamods]{import-file}.
Removes everything but numeric variables (i.e. open ends) and adds
no_weighting variable.
}
\details{
Intended for use with \link[datamods]{import-file}, but might work with other
loading methods / source file types.
}
