% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionVars.R
\name{selectionVars}
\alias{selectionVars}
\alias{selectionVars_server}
\alias{rowChoice_ui}
\alias{rowChoice_server}
\alias{colChoice_ui}
\alias{colChoice_server}
\title{Shiny modules for specifying row and column variables}
\usage{
selectionVars_server(id, varLabels, exclude)

rowChoice_ui(id)

rowChoice_server(id, selection_vars, varLabels)

colChoice_ui(id)

colChoice_server(id, selection_vars, varLabels)
}
\arguments{
\item{id}{Module ID}

\item{varLabels}{Dataframe (e.g. as generated by
\link[tveDataLoader]{get_varLabels}) containing two columns:
\itemize{
\item \code{variable}: the variable names
\item \code{label}: the variable labels
}}

\item{exclude}{character vector of variable names to exclude}

\item{selection_vars}{Reactive character vector of the variable labels from
which users can select row / column breaks. Can be generated by
\code{selectionVars_server} or input manually.}
}
\value{
\itemize{
\item \code{selectionVars_server}: Character vector of the variable labels that can
be chosen. Intended to be stored as a reactive then passed to \code{rowChoice} or
\code{colChoice} in the instance where the same set of variables can be used for
both.
\item \code{rowChoice} / \code{colChoice}:
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item server: Character vector of the variable labels to be used as row /
column breaks respectively.
}
}
}
\description{
\itemize{
\item \code{selectionVars_server}: Takes a variable labels dataframe, filters out
variables that should not be used for row / column breaks in the output
table, then \link[dplyr]{pull}s the variable labels, ready for passing into
\code{rowChoice} or \code{colChoice}. Intended to be used for convenience when the
same variables can be passed to both. No UI element.
\item \code{rowChoice} allows specification of multiple variables to be
used as rows in the output table.
\item \code{colChoice} restricts selection to a single variable to be used as column
breaks in the output table.
}
}
