% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table_unlabelled.R
\name{single_calcs_unlabs}
\alias{single_calcs_unlabs}
\title{calculate table for single code questions}
\usage{
single_calcs_unlabs(dat, col_var, weight_var, flag_list)
}
\arguments{
\item{dat}{data}

\item{col_var}{Character string containing the variables to be used as
columns in the output table.}

\item{weight_var}{Character string containing the variable name of the
weight to be used.}

\item{flag_list}{named list of character vectors for category
As generated by \link[tveTableTool]{identify_var_type}}
}
\value{
table
}
\description{
calculate table for single code questions
}
\examples{
''


}
