% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_list}
\alias{test_list}
\title{Test data for variable classification.}
\format{
An object of class \code{list} of length 5.
}
\usage{
test_list
}
\description{
Output from tveDataLoader::load_sav containing a list of tibbles with the following components:
}
\details{
\itemize{
\item \code{loaded_data}: A tibble containing the raw data with zapped variables and tidy variable names
\item \code{var_labels}: A tibble containing variable labels with columns:
\itemize{
\item \code{var_name}: Name of the variable
\item \code{var_label}: Label for the variable
}
\item \code{val_labels}: A tibble containing value labels with columns:
\itemize{
\item \code{var_name}: Name of the variable
\item \code{val_label}: Label for the value
\item \code{val_value}: Numeric value corresponding to the label
}
\item \code{no_var_labels}: A tibble containing variables with no variable labels:
\itemize{
\item \code{var_name}: Name of the variable
}
\item \code{no_val_labels}: A tibble containing variables with no value labels:
\itemize{
\item \code{var_name}: Name of the variable
}
}
}
\keyword{datasets}
