% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testTableData}
\alias{testTableData}
\title{Test data for generating tables.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 750 rows and 208 columns.
}
\usage{
testTableData
}
\description{
Has 3 example weight variables:
}
\details{
\itemize{
\item \code{no_weight}: all records have a value of 1 so no weighting is applied
\item \code{gender_weight1}: Male weighted to 0.5, female weighted to 1.5, other
weighted to 1
\item \code{gender_weight2}: Male weighted to 1.5, female weighted to 0.5, other
weighted to 1
}
}
\keyword{datasets}
