% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tveTableTool-package.R
\docType{package}
\name{tveTableTool-package}
\alias{tveTableTool}
\alias{tveTableTool-package}
\title{tveTableTool: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: first name last name \email{email@email.com}

}
\keyword{internal}
