% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightVariable.R
\name{weightVariable}
\alias{weightVariable}
\alias{weightVariable_ui}
\alias{weightVariable_server}
\title{Shiny module for specifying a weighting variable}
\usage{
weightVariable_ui(id)

weightVariable_server(id, varLabels)
}
\arguments{
\item{id}{Module ID}

\item{varLabels}{Dataframe (e.g. as generated by
\link[tveDataLoader]{get_varLabels}) containing two columns:
\itemize{
\item \code{variable}: the variable names
\item \code{label}: the variable labels
}}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: A string of either the weight variables name or "no_weighting" if
no weight is selected
}
}
\description{
Generates an input switch for whether a weight variable should
be used, and if set to yes, a picker input listing all variables from the
data apart from the first (which is assumed to be respid).
}
