% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{calculate_freqs}
\alias{calculate_freqs}
\title{Calculate weighted frequencies and proportions}
\usage{
calculate_freqs(
  input_data,
  col_var,
  respid_var,
  weight_var,
  var_labels,
  val_labels
)
}
\description{
Calculate weighted frequencies and proportions
}
