% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{classify_variables}
\alias{classify_variables}
\title{Classify variables by type}
\usage{
classify_variables(
  loaded_data,
  val_labels,
  no_val_labels = NULL,
  binary_labels = c(No = 0, Yes = 1),
  meta_vars
)
}
\description{
Classify variables by type
}
