% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{create_tabulations}
\alias{create_tabulations}
\title{Create all tabulations}
\usage{
create_tabulations(
  loaded_data,
  var_labels,
  val_labels,
  no_val_labels = NULL,
  respid_var = "respid",
  col_var = NULL,
  weight_var = NULL
)
}
\description{
Create all tabulations
}
