% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{generate_table}
\alias{generate_table}
\title{wraps the 3 generate table functions into one to produce final table}
\usage{
generate_table(
  dat,
  row_vars,
  col_var,
  weight_var,
  variable_labels,
  value_labels
)
}
\arguments{
\item{dat}{data}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{col_var}{Character string containing the variables to be used as
columns in the output table.}

\item{weight_var}{Character string containing the variable name of the
weight to be used.}

\item{variable_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item label - the variable labels of \code{dat}
As generated by \link[tveDataLoader]{get_varLabels}
}}

\item{value_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item value - the values \code{dat}
\item value label - the value labels of \code{dat}
As generated by \link[tveDataLoader]{get_valLabels}
}}
}
\value{
final output table
}
\description{
wraps the 3 generate table functions into one to produce final table
}
\examples{
''

}
