% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{generate_table}
\alias{generate_table}
\title{Generate frequency / proportions table from survey data}
\usage{
generate_table(dat, row_vars, col_var, variable_labels, percents = FALSE)
}
\arguments{
\item{dat}{A dataframe containing the variables to be used as rows and
columns in the output table.}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{col_var}{Character string containing the variables ot be used as
columns in the output table.}

\item{variable_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item label - the variable labels of \code{dat}
As generated by \link[tveDataLoader]{get_varLabels}
}}

\item{perecents}{TRUE if column percents required, FALSE If counts required.}
}
\description{
Generate frequency / proportions table from survey data
}
