% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{generate_table}
\alias{generate_table}
\title{Generate frequency / proportions table from survey data}
\usage{
generate_table(
  dat,
  row_vars,
  col_var,
  weight_var,
  variable_labels,
  percents = "columns"
)
}
\arguments{
\item{dat}{A dataframe containing the variables to be used as rows and
columns in the output table.}

\item{row_vars}{Character vector containing the variables to be used as rows
in the table.}

\item{col_var}{Character string containing the variables to be used as
columns in the output table.}

\item{weight_var}{Character string containing the variable name of the
weight to be used.}

\item{variable_labels}{A dataframe consisting of 2 columns:
\enumerate{
\item variable - the variable names of \code{dat}
\item label - the variable labels of \code{dat}
As generated by \link[tveDataLoader]{get_varLabels}
}}

\item{percents}{Either "columns" for column percents, "rows" for row
percents or "none" for counts.}
}
\value{
Tibble containing \code{Variable} name, variable \code{Label}, variable
\code{Value}, a \code{Total} column and a column for each value within the specified
column variable.
}
\description{
Generates a frequency (counts) or proportions (row or column percents) table
from a dataframe. Multiple row variables can be specified, but only one
column variable.
}
\details{
Respondent weighting can be applied by specifying a variable within the data
to weight by. Note, no sense checks are applied to the weight, so results
could be extreme is not used properly.
}
