% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tveTableTool-package.R
\docType{package}
\name{tveTableTool-package}
\alias{tveTableTool}
\alias{tveTableTool-package}
\title{tveTableTool: DTAG Analytics Cross-tabulation Services}
\description{
This package provides all functionality to deliver cross-tabulation as per DTAG Analytics standards. It includes all tabulation functions, configurations, and a support Shiny application for interactive use of these.
}
\author{
\strong{Maintainer}: Juan Hernandez \email{email@email.com}

}
\keyword{internal}
